requirements.txt
streamlit
transformers
torch
PyMuPDF
sentence-transformers
faiss-cpu
